Server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
chromadb==0.6.3
python-dotenv==1.0.1